Java zip file execution strategy and refactorings

+++ Open Judge System/Workers/OJS.Workers.Executors/StandardProcessExecutor.cs
@@ -44,9 +44,10 @@
             {
                 if (process == null)
                 {
-                    var errorMessage = string.Format("Could not start process: {0}!", fileName);
-                    throw new Exception(errorMessage);
+                    throw new Exception($"Could not start process: {fileName}!");
                 }
+
+                process.PriorityClass = ProcessPriorityClass.High;
 
                 // Write to standard input using another thread
                 process.StandardInput.WriteLineAsync(inputData).ContinueWith(
@@ -83,6 +84,12 @@
                         while (true)
                         {
                             // ReSharper disable once AccessToDisposedClosure
+                            if (process.HasExited)
+                            {
+                                return;
+                            }
+
+                            // ReSharper disable once AccessToDisposedClosure
                             var peakWorkingSetSize = process.PeakWorkingSet64;
 
                             result.MemoryUsed = Math.Max(result.MemoryUsed, peakWorkingSetSize);
@@ -102,7 +109,15 @@
                 var exited = process.WaitForExit((int)(timeLimit * 1.5));
                 if (!exited)
                 {
-                    process.Kill();
+                    // Double check if the process has exited before killing it
+                    if (!process.HasExited)
+                    {
+                        process.Kill();
+
+                        // Approach: https://msdn.microsoft.com/en-us/library/system.diagnostics.process.kill(v=vs.110).aspx#Anchor_2
+                        process.WaitForExit();
+                    }
+
                     result.Type = ProcessExecutionResultType.TimeLimit;
                 }
